+++ config/pricing.php
@@ -1,0 +1,10 @@
+<?php
+
+return [
+    'pricing_price_1' => env('PRICING_PROMO_1', 1),
+    'pricing_price_2' => env('PRICING_PROMO_2', 5),
+    'pricing_price_3' => env('PRICING_PROMO_3', 10),
+    'pricing_price_4' => env('PRICING_PROMO_4', 20),
+    'pricing_price_5' => env('PRICING_PROMO_5', 50),
+    'pricing_price_6' => env('PRICING_PROMO_6', 100),
+];