+++ resources/views/blog/show.blade.php
@@ -17,6 +17,16 @@
 </div>
 <!-- Page Header End -->
 
+<ol class="breadcrumb">
+    <li><a href="{{ config('app.url') }}">Inicio</a></li>
+    <li class="ml-2">/</li>
+    <li class="ml-2"><a href="{{ route('blog_index') }}">Blog de Noticias Comercio y Compra venta en Cuba</a></li>
+    <li class="ml-2">/</li>
+    <li class="ml-2"><a href="{{ route('blog_index', ['blog_category_slug' => $blog_post->category->slug]) }}">{{ $blog_post->category->name }}</a></li>
+    <li class="ml-2">/</li>
+    <li class="ml-2">{{ $blog_post->title }}</li>
+</ol>
+
 <!-- Start Content -->
 <div id="content" class="section-padding">
     <div class="container">
@@ -56,10 +66,13 @@
                     </div>
                     <!-- Post Content -->
                 </div>
-
-                @include('blog.author-bio')
-
                 <!-- End Post -->
+
+                <!-- Author Bio -->
+                @include('blog.author-bio')
+                <!-- End Author Bio -->
+
+
                 <div id="disqus_thread"></div>
 
             </div>